+++ src/NetAppFiles/NetAppFiles/BackupPolicy/GetNetAppFilesBackupPolicy.cs
@@ -23,6 +23,7 @@ using System.Globalization;
 using Microsoft.Azure.Commands.NetAppFiles.Helpers;
 using System.Linq;
 using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
+using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 
 namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
 {
@@ -30,6 +31,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
         "Get",
         ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "NetAppFilesBackupPolicy",
         DefaultParameterSetName = FieldsParameterSet), OutputType(typeof(PSNetAppFilesBackupPolicy))]
+    [CmdletOutputBreakingChange(typeof(PSNetAppFilesBackupPolicy), DeprecatedOutputProperties = new string[] { "YearlyBackupsToKeep" })]
     [Alias("Get-AnfBackupPolicy")]
     public class GetAzureRmNetAppFilesBackupPolicy : AzureNetAppFilesCmdletBase
     {