+++ src/Storage/Storage.Management/Blob/AddAzureStorageContainerLegalHold.cs
@@ -91,7 +91,21 @@ namespace Microsoft.Azure.Commands.Management.Storage
         [Parameter(HelpMessage = "Container LegalHold Tags", Mandatory = true)]
         [ValidateNotNullOrEmpty]
         public string[] Tag { get; set; }
-        
+
+        [Parameter(Mandatory = false, HelpMessage = "When enabled, new blocks can be written to both 'Appened and Block Blobs' while maintaining immutability protection and compliance. Only new blocks can be added and any existing blocks cannot be modified or deleted.")]
+        public bool AllowProtectedAppendWriteAll
+        {
+            get
+            {
+                return allowProtectedAppendWriteAll is null ? false : allowProtectedAppendWriteAll.Value;
+            }
+            set
+            {
+                allowProtectedAppendWriteAll = value;
+            }
+        }
+        private bool? allowProtectedAppendWriteAll;
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -116,7 +130,8 @@ namespace Microsoft.Azure.Commands.Management.Storage
                         this.ResourceGroupName,
                         this.StorageAccountName,
                         this.Name,
-                        new List<string>(this.Tag));
+                        new List<string>(this.Tag),
+                        this.allowProtectedAppendWriteAll);
                 WriteObject(new PSLegalHold(legalHold));
             }
         }